Fix texture_id type

+++ windows/video_outlet.hpp
@@ -1,13 +1,14 @@
-#include <mutex>
 #include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 #include <flutter/standard_method_codec.h>
+
+#include <mutex>
 
 class VideoOutlet {
  public:
   VideoOutlet(flutter::TextureRegistrar* texture_registrar);
 
-  int32_t texture_id() const { return texture_id_; }
+  int64_t texture_id() const { return texture_id_; }
 
   void OnVideo(uint8_t* buffer, int32_t width, int32_t height);
 
@@ -17,6 +18,6 @@ class VideoOutlet {
   FlutterDesktopPixelBuffer flutter_pixel_buffer_;
   flutter::TextureRegistrar* texture_registrar_ = nullptr;
   std::unique_ptr<flutter::TextureVariant> texture_ = nullptr;
-  int32_t texture_id_;
+  int64_t texture_id_;
   mutable std::mutex mutex_;
 };
